+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/ConfirmationRegistry.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.paymentsheet.confirmation
+package com.stripe.android.paymentelement.confirmation
 
 import androidx.lifecycle.SavedStateHandle
