+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/IntentConfirmationDefinition.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.paymentsheet.confirmation
+package com.stripe.android.paymentelement.confirmation
 
 import androidx.activity.result.ActivityResultCaller
 import androidx.activity.result.ActivityResultLauncher
