+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/IntentConfirmationInterceptorKtx.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.paymentsheet.confirmation
+package com.stripe.android.paymentelement.confirmation
 
 import com.stripe.android.paymentsheet.addresselement.toConfirmPaymentIntentShipping
