+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/analytics/EventReporter.kt
@@ -4,8 +4,8 @@ import androidx.annotation.Keep
 import com.stripe.android.model.CardBrand
 import com.stripe.android.model.LinkMode
 import com.stripe.android.model.PaymentMethodCode
+import com.stripe.android.paymentelement.confirmation.DeferredIntentConfirmationType
 import com.stripe.android.paymentsheet.PaymentSheet
-import com.stripe.android.paymentsheet.confirmation.DeferredIntentConfirmationType
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.state.PaymentElementLoader
