+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/paymentdatacollection/bacs/BacsMandateDataTest.kt
@@ -3,8 +3,8 @@ package com.stripe.android.paymentsheet.paymentdatacollection.bacs
 import com.google.common.truth.Truth.assertThat
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCreateParams
+import com.stripe.android.paymentelement.confirmation.ConfirmationHandler
 import com.stripe.android.paymentsheet.PaymentSheetFixtures
-import com.stripe.android.paymentsheet.confirmation.ConfirmationHandler
 import com.stripe.android.paymentsheet.state.PaymentElementLoader
 import org.junit.Test
